Add create-calling flow and shareable action checklist
- + button on Ward Callings and drill-down: pick a group, tap + after a
  calling, name it — inserted halfway in display order (now fractional),
  vacant, and shown in orange until an LCR import confirms it. Pending
  callings survive import slot replacement; a matching import row clears
  the flag and hands over to the PDF slot.
- Share icon on Open Callings copies a markdown action checklist grouped
  by assignee (release/call/select-candidate items) plus a Ward Clerk
  section listing callings still to be added in LCR.

+++ Callings/Services/CallingSeedRules.swift
@@ -7,9 +7,9 @@ enum CallingSeedRules {
 
     /// Display order from calling-name keywords: President first, then
     /// counselors, secretary, and the rest, matching the spec's required order.
-    static func displayOrder(for name: String) -> Int {
+    static func displayOrder(for name: String) -> Double {
         let lowered = name.lowercased()
-        let rules: [(keyword: String, order: Int)] = [
+        let rules: [(keyword: String, order: Double)] = [
             ("bishop", 0), ("president", 0),
             ("first counselor", 10), ("first assistant", 10),
             ("second counselor", 20), ("second assistant", 20),
